9/11 유기동물 집어넣을것만 넣음, 검색기능 후 상품리스트 경로설정, 1등으로 커밋함

+++ src/main/java/com/springboot/meongnyang_Jiphapso/dto/StrayAnimalDto.java
@@ -1,7 +1,6 @@
 package com.springboot.meongnyang_Jiphapso.dto;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.time.LocalDate;
 
 import org.springframework.web.multipart.MultipartFile;
@@ -10,7 +9,7 @@ import lombok.Data;
 
 @Data
 public class StrayAnimalDto {
-	 private BigInteger stray_no;
+	 private Long stray_no;
 	 private String stray_category;
 	 private String stray_name;
 	 private String stray_age;
